Renames submodule file
