+++ plugboard/utils/__init__.py
@@ -1,7 +1,7 @@
 """Provides utility functions for use throughout the code."""
 
-from plugboard.utils.as_dict_mixin import Exportable, ExportMixin
 from plugboard.utils.entities import EntityIdGen
+from plugboard.utils.export_mixin import Exportable, ExportMixin
 from plugboard.utils.random import gen_rand_str
 from plugboard.utils.registry import ClassRegistry
